Обновил парам тесты

+++ src/main/java/pro/sky/skyprocalculator/service/CalculatorService.java
@@ -5,5 +5,5 @@ public interface CalculatorService {
     Integer sum(Integer num1, Integer num2);
     Integer minus(Integer num1, Integer num2);
     Integer multiply(Integer num1, Integer num2);
-    Double divide(Integer num1, Integer num2);
+    Integer divide(Integer num1, Integer num2);
 }
