+++ src/main/java/pro/sky/skyprocalculator/service/CalculatorServiceImpl.java
@@ -36,14 +36,14 @@ public class CalculatorServiceImpl implements CalculatorService {
     }
 
     @Override
-    public Double divide(Integer num1, Integer num2) {
+    public Integer divide(Integer num1, Integer num2) {
         if (num1 == null && num2 == null) {
             throw new NullPointerException("Нет входных значений");
         }
         if (num2 == 0) {
             throw new DivisionByZero("Делить на 0 нельзя");
         }
-        return (double) (num1 / num2);
+        return num1 / num2;
 
     }
 }
